+++ db/migrate/20180803174948_create_capitalizations.rb
@@ -1,8 +1,8 @@
 class CreateCapitalizations < ActiveRecord::Migration[5.2]
   def change
     create_table :capitalizations do |t|
-      t.integer :rate
-      
+      t.decimal :rate
+
       t.timestamps
     end
   end
